+++ src/silent-lint/silent-lint.class.ts
@@ -1,4 +1,4 @@
-import { IZLinter } from '../lint/linter.interface';
+import { IZLinter } from '../lint-janitor/linter.interface';
 
 /**
  * Represents a silent linter that always succeeds.